Make pass all old test (1 new fail) with git2 lib
